Abstracted common functionality from PollableResult and SubscribableResult into a separate base class
