+++ civis/polling.py
@@ -1,28 +1,14 @@
 from concurrent import futures
 import time
 
-from civis.base import CivisJobFailure
+from civis.base import CivisJobFailure, CivisAsyncResultBase, FAILED, DONE
 from civis.response import Response
 
 
-FINISHED = ['success', 'succeeded']
-FAILED = ['failed']
-NOT_FINISHED = ['queued', 'running']
-CANCELLED = ['cancelled']
-DONE = FINISHED + FAILED + CANCELLED
 _DEFAULT_POLLING_INTERVAL = 15
 
-# Translate Civis state strings into `future` state strings
-STATE_TRANS = {}
-for name in FINISHED + FAILED:
-    STATE_TRANS[name] = futures._base.FINISHED
-for name in NOT_FINISHED:
-    STATE_TRANS[name] = futures._base.RUNNING
-for name in CANCELLED:
-    STATE_TRANS[name] = futures._base.CANCELLED_AND_NOTIFIED
 
-
-class PollableResult(futures.Future):
+class PollableResult(CivisAsyncResultBase):
     """A class for tracking pollable results.
 
     This class will begin polling immediately upon creation, and poll for
@@ -81,42 +67,6 @@ class PollableResult(futures.Future):
         self._last_result = None
 
         self._self_polling_executor = None
-
-    def __repr__(self):
-        # Almost the same as the superclass's __repr__, except we use
-        # the `_civis_state` rather than the `_state`.
-        with self._condition:
-            if self._civis_state in FINISHED + FAILED:
-                if self.exception():
-                    return '<%s at %#x state=%s raised %s>' % (
-                        self.__class__.__name__,
-                        id(self),
-                        self._civis_state,
-                        self._exception.__class__.__name__)
-                else:
-                    return '<%s at %#x state=%s returned %s>' % (
-                        self.__class__.__name__,
-                        id(self),
-                        self._civis_state,
-                        self.result().__class__.__name__)
-            out = '<%s at %#x state=%s>' % (self.__class__.__name__,
-                                            id(self),
-                                            self._civis_state)
-            return out
-
-    def cancel(self):
-        """Not currently implemented."""
-        raise NotImplementedError("Running jobs cannot currently be cancelled")
-
-    def succeeded(self):
-        """Return ``True`` if the job completed in Civis with no error."""
-        with self._condition:
-            return self._civis_state in FINISHED
-
-    def failed(self):
-        """Return ``True`` if the Civis job failed."""
-        with self._condition:
-            return self._civis_state in FAILED
 
     def _wait_for_completion(self):
         """Poll the job every `polling_interval` seconds. Blocks until the
@@ -186,20 +136,3 @@ class PollableResult(futures.Future):
                         self.set_result(self._last_result)
 
             return self._last_result
-
-    @property
-    def _civis_state(self):
-        """State as returned from Civis."""
-        with self._condition:
-            return self._check_result().state
-
-    @property
-    def _state(self):
-        """State of the PollableResult in `future` language."""
-        with self._condition:
-            return STATE_TRANS[self._civis_state]
-
-    @_state.setter
-    def _state(self, value):
-        # Ignore attempts to set the _state from the `Future` superclass
-        pass
