+++ civis/pubnub.py
@@ -1,4 +1,4 @@
-from concurrent import futures
+from civis.base import CivisJobFailure, CivisAsyncResultBase, FAILED, DONE
 
 try:
     from pubnub.pubnub import PubNub
@@ -7,9 +7,6 @@ try:
     has_pubnub = True
 except ImportError:
     has_pubnub = False
-
-from civis.base import CivisJobFailure
-from civis.polling import FINISHED, FAILED, DONE, STATE_TRANS
 
 
 if has_pubnub:
@@ -34,7 +31,7 @@ if has_pubnub:
             pass
 
 
-class SubscribableResult(futures.Future):
+class SubscribableResult(CivisAsyncResultBase):
     def __init__(self, poller, poller_args, client):
         super().__init__()
 
@@ -42,29 +39,6 @@ class SubscribableResult(futures.Future):
         self.poller_args = poller_args
         self._client = client
         self._pubnub = self._subscribe()
-
-    def __repr__(self):
-        # Almost the same as the superclass's __repr__, except we use
-        # the `_civis_state` rather than the `_state`.
-        with self._condition:
-            if self._civis_state in FINISHED + FAILED:
-                if self.exception():
-                    return '<%s at %#x state=%s raised %s>' % (
-                        self.__class__.__name__,
-                        id(self),
-                        self._civis_state,
-                        self._exception.__class__.__name__)
-                else:
-                    return '<%s at %#x state=%s returned %s>' % (
-                        self.__class__.__name__,
-                        id(self),
-                        self._civis_state,
-                        self.result().__class__.__name__)
-            out = '<%s at %#x state=%s>' % (self.__class__.__name__,
-                                            id(self),
-                                            self._civis_state)
-            return out
-
 
     def _subscribe(self):
         pnconfig, channels = self._pubnub_config()
@@ -86,7 +60,6 @@ class SubscribableResult(futures.Future):
         pnconfig.reconnect_policy = True
         return pnconfig, channels
 
-
     def _check_api_result(self, result=None):
         with self._condition:
             if result is None:
@@ -105,42 +78,3 @@ class SubscribableResult(futures.Future):
                 if self._pubnub:
                     self._pubnub.unsubscribe_all()
                 self.set_result(result)
-
-
-    def cancel(self):
-        """Not currently implemented."""
-        raise NotImplementedError("Running jobs cannot currently be cancelled")
-
-    def succeeded(self):
-        """Return ``True`` if the job completed in Civis with no error."""
-        with self._condition:
-            return self._civis_state in FINISHED
-
-    def failed(self):
-        """Return ``True`` if the Civis job failed."""
-        with self._condition:
-            return self._civis_state in FAILED
-
-    def _check_result(self):
-        with self._condition:
-            if self._result is not None:
-                return self._result
-
-    @property
-    def _civis_state(self):
-        """State as returned from Civis."""
-        with self._condition:
-            if self._check_result():
-                return self._check_result().state
-            return 'running'
-
-    @property
-    def _state(self):
-        """State of the PollableResult in `future` language."""
-        with self._condition:
-            return STATE_TRANS[self._civis_state]
-
-    @_state.setter
-    def _state(self, value):
-        # Ignore attempts to set the _state from the `Future` superclass
-        pass
